requirements.txt
streamlit
faiss-cpu
numpy
requests
python-dotenv